--- packages/button-link/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "@equinor/fusion-wc-link-button",
-  "version": "0.0.0",
-  "description": "",
-  "main": "lib/index.js",
-  "customElements": "lib/custom-elements.json",
-  "scripts": {
-    "build": "tsc -b",
-    "postbuild": "npm run typedoc",
-    "typedoc": "cem analyze --litelement --globs \"src/**/*.ts\" --outdir lib",
-    "prepack": "npm run build && fwc-build"
-  },
-  "keywords": [
-    "equinor",
-    "fusion",
-    "web components",
-    "button",
-    "link",
-    "url",
-    "href"
-  ],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@equinor/fusion-wc-button": "^1.1.34",
-    "@equinor/fusion-wc-core": "^1.0.5",
-    "@equinor/fusion-wc-icon": "^1.0.27",
-    "@material/mwc-button": "^0.27.0",
-    "lit": "2.4.1"
-  },
-  "devDependencies": {
-    "@custom-elements-manifest/analyzer": "^0.6.6",
-    "@equinor/fusion-wc-builder": "^2.0.0",
-    "typescript": "^4.8.4"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/equinor/fusion-web-components.git",
-    "directory": "packages/button-link"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
