--- packages/button-link/src/element.css.ts
@@ -1,13 +1,0 @@
-import { css } from 'lit';
-// import { styles as theme } from '@equinor/fusion-web-theme';
-
-import { styles as buttonStyle } from '@equinor/fusion-wc-button/src/element.css';
-
-export const style = css`
-  :host {
-  }
-`;
-
-export const styles = [buttonStyle, style];
-
-export default styles;
