--- packages/button-link/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "./lib"
-  },
-  "include": [
-    "src/**/*",
-    "types/*.d.ts"
-  ],
-  "references": [
-    {
-      "path": "../button"
-    },
-    {
-      "path": "../core"
-    },
-    {
-      "path": "../icon"
-    }
-  ]
-}
