+++ packages/button/package.json
@@ -3,6 +3,27 @@
   "version": "1.1.34",
   "description": "",
   "main": "lib/index.js",
+  "exports": {
+    ".": "./lib/index.js",
+    "./README.button.md": "./README.button.md",
+    "./README.link-button.md": "./README.link-button.md",
+    "./button": "./lib/button/index.js",
+    "./link-button": "./lib/link-button/index.js"
+  },
+  "types": "lib/index.d.ts",
+  "typesVersions": {
+    ">=4.2": {
+      ".": [
+        "lib/index.d.ts"
+      ],
+      "button": [
+        "lib/button/index.d.ts"
+      ],
+      "link-button": [
+        "lib/link-button/index.d.ts"
+      ]
+    }
+  },
   "customElements": "lib/custom-elements.json",
   "scripts": {
     "build": "tsc -b",
