+++ packages/button/src/button/index.ts
@@ -1,0 +1,23 @@
+import { fusionElement } from '@equinor/fusion-wc-core';
+import ButtonElement, { ButtonElementProps } from './element';
+export * from './element';
+
+export const tag = 'fwc-button';
+
+@fusionElement(tag)
+export default class _ extends ButtonElement {}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    [tag]: ButtonElement;
+  }
+
+  namespace JSX {
+    interface IntrinsicElements {
+      [tag]: React.DetailedHTMLProps<
+        React.PropsWithChildren<ButtonElementProps & React.HTMLAttributes<ButtonElement>>,
+        ButtonElement
+      >;
+    }
+  }
+}
