+++ packages/button/src/link-button/element.css.ts
@@ -1,0 +1,11 @@
+import { css } from 'lit';
+import { styles as buttonStyle } from '../button/element.css';
+
+export const style = css`
+  :host {
+  }
+`;
+
+export const styles = [buttonStyle, style];
+
+export default styles;
