+++ storybook/package.json
@@ -16,7 +16,6 @@
     "@equinor/fusion-wc-avatar": "^1.0.39",
     "@equinor/fusion-wc-badge": "^0.2.28",
     "@equinor/fusion-wc-button": "^1.1.34",
-    "@equinor/fusion-wc-link-button": "^0.0.0",
     "@equinor/fusion-wc-checkbox": "^0.3.12",
     "@equinor/fusion-wc-chip": "^0.2.19",
     "@equinor/fusion-wc-date": "^0.2.24",
