+++ storybook/stories/button-link.stories.mdx
@@ -1,8 +1,8 @@
 import { Meta, Story, Canvas, ArgsTable, Description } from '@storybook/addon-docs';
 
 import { iconNames } from '@equinor/fusion-wc-icon';
-import { LinkTarget } from '@equinor/fusion-wc-link-button';
-import readme from '@equinor/fusion-wc-button/README.md';
+import { LinkTarget } from '@equinor/fusion-wc-button/link-button';
+import readme from '@equinor/fusion-wc-button/README.link-button.md';
 
 import LinkButton from './Components/ButtonLink';
