+++ storybook/stories/button.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Canvas, ArgsTable, Description } from '@storybook/addon-docs';
 
 import { iconNames } from '@equinor/fusion-wc-icon';
-import readme from '@equinor/fusion-wc-button/README.md';
+import readme from '@equinor/fusion-wc-button/README.button.md';
 
 import Button from './Components/Button';
