+++ tsconfig.json
@@ -24,9 +24,6 @@
       "path": "./packages/button"
     },
     {
-      "path": "./packages/button-link"
-    },
-    {
       "path": "./packages/checkbox"
     },
     {
